Refactor the MaterialXRenderHw folder for consistency with other libraries
