+++ source/MaterialXRenderHw/WindowCocoaWrappers.h
