+++ source/MaterialXRenderHw/WindowWrapper.h
@@ -7,6 +7,7 @@
 #define MATERIALX_WINDOWWRAPPER_H
 
 #if defined(_WIN32)
+#define WIN32_LEAN_AND_MEAN
 #include <windows.h>
 #elif defined(__linux__)
 #include <X11/X.h> // for Window
